Consolidate test suites for viewset. Fix polling time for 0th place in the queue.
